requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
Deprecated==1.2.13
distlib==0.3.6
Django==3.2.16
djangorestframework==3.14.0
exceptiongroup==1.1.0
filelock==3.8.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
lxml==4.9.2
outcome==1.2.0
packaging==21.3
pdfdocument==4.0.0
Pillow==9.3.0
platformdirs==2.5.2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
selenium==4.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.0.7
zipp==3.9.0
zope.interface==5.5.2
webdriver_manager